Fix long to int

+++ src/main/java/com/sharif/thunder/BotConfig.java
@@ -52,7 +52,7 @@ public class BotConfig {
   @Getter(AccessLevel.NONE)
   private boolean npImages;
   @Getter
-  private long port;
+  private int port;
   private long ownerId;
   private long maxSeconds;
 
@@ -90,7 +90,7 @@ public class BotConfig {
       dbPass = config.getString("dbpass");
       defaultLyrics = config.getString("lyrics.default");
       emiliaKey = config.getString("emiliakey");
-      port = config.getLong("port");
+      port = config.getInt("port");
     } catch (Exception ex) {
       System.out.println(ex + ": " + ex.getMessage() + "\n\nConfig Location: " + path.toAbsolutePath().toString());
     }
